# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  auth.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.LogoutInput:
    properties:
      refresh_token:
        type: string
    type: object
  auth.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.RenewInput:
    properties:
      refresh_token:
        type: string
    type: object
  auth.TokenPair:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  spending.CreateSpendingInput:
    properties:
      category:
        type: string
      name:
        type: string
      observation:
        type: string
      price:
        type: number
      spent_at:
        example: "2026-03-15T00:00:00Z"
        type: string
    type: object
  spending.Spending:
    properties:
      category:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      observation:
        type: string
      order_number:
        type: integer
      price:
        type: number
      spent_at:
        example: "2026-03-15T00:00:00Z"
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  spending.UpdateSpendingInput:
    properties:
      category:
        type: string
      name:
        type: string
      observation:
        type: string
      price:
        type: number
      spent_at:
        description: SpentAt must be in RFC3339 format (e.g. 2026-03-15T00:00:00Z).
        example: "2026-03-15T00:00:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: SaaS finance management API.
  title: CoFi Finance API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password. Returns a JWT access
        token (2h) and a refresh token (7d).
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenPair'
        "400":
          description: Missing or malformed request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the current access token and optionally a refresh token.
        Both tokens are blacklisted in Redis until they naturally expire.
      parameters:
      - description: Optional refresh token to revoke
        in: body
        name: body
        schema:
          $ref: '#/definitions/auth.LogoutInput'
      produces:
      - application/json
      responses:
        "204":
          description: Logged out successfully
        "401":
          description: Missing, invalid, or blacklisted access token; or invalid refresh
            token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password. A verification
        email is sent to the provided address. The account must be verified before
        sign-in is allowed.
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Account created. Verification email sent.
        "400":
          description: Missing, malformed, or invalid fields
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register
      tags:
      - auth
  /api/auth/renew:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access token (2h) and
        refresh token (7d).
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RenewInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenPair'
        "400":
          description: Missing or malformed request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Renew tokens
      tags:
      - auth
  /api/auth/verify-email:
    get:
      description: Confirm a user's email address using the token from the verification
        email. The token is single-use and expires in 24 hours. On success, redirects
        to the frontend /email-verified page.
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "303":
          description: Redirect to frontend /email-verified page
        "400":
          description: Missing, invalid, expired, or already-used token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verify email
      tags:
      - auth
  /api/spendings:
    get:
      description: 'Returns all spendings for the authenticated user, ordered by most
        recent first. Optionally filter by creation date using ''from'' and/or ''to''
        query parameters (format: YYYY-MM-DD). The ''to'' date is inclusive (includes
        the full day).'
      parameters:
      - description: Filter spendings created on or after this date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Filter spendings created on or before this date, inclusive (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/spending.Spending'
            type: array
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List spendings
      tags:
      - spending
    post:
      consumes:
      - application/json
      description: Creates a new spending for the authenticated user.
      parameters:
      - description: Spending data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/spending.CreateSpendingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/spending.Spending'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create spending
      tags:
      - spending
  /api/spendings/{id}:
    delete:
      description: Deletes the spending with the given ID if it belongs to the authenticated
        user.
      parameters:
      - description: Spending ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Spending not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete spending
      tags:
      - spending
    patch:
      consumes:
      - application/json
      description: Updates the spending with the given ID if it belongs to the authenticated
        user. The spent_at field must be in RFC3339 format (e.g. 2026-03-15T00:00:00Z).
      parameters:
      - description: Spending ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated spending data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/spending.UpdateSpendingInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/spending.Spending'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Spending not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update spending
      tags:
      - spending
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"